+++ removelineagecommand.cpp
@@ -556,7 +556,7 @@ int RemoveLineageCommand::readTax(){
 		bool wroteSomething = false;
 		
 		bool taxonsHasConfidence = false;
-		vector< map<string, int> > searchTaxons;
+		vector< map<string, float> > searchTaxons;
 		string noConfidenceTaxons = taxons;
 		int hasConPos = taxons.find_first_of('(');
 		if (hasConPos != string::npos) {  
@@ -618,7 +618,7 @@ int RemoveLineageCommand::readTax(){
 					if (pos != string::npos) { //if yes, then are the confidences okay
 						
 						bool remove = false;
-						vector< map<string, int> > usersTaxon = getTaxons(newtax);
+						vector< map<string, float> > usersTaxon = getTaxons(newtax);
 						
 						//the usersTaxon is most likely longer than the searchTaxons, and searchTaxon[0] may relate to userTaxon[4]
 						//we want to "line them up", so we will find the the index where the searchstring starts
@@ -686,10 +686,10 @@ int RemoveLineageCommand::readTax(){
 	}
 }
 /**************************************************************************************************/
-vector< map<string, int> > RemoveLineageCommand::getTaxons(string tax) {
+vector< map<string, float> > RemoveLineageCommand::getTaxons(string tax) {
 	try {
 		
-		vector< map<string, int> > t;
+		vector< map<string, float> > t;
 		string taxon = "";
 		int taxLength = tax.length();
 		for(int i=0;i<taxLength;i++){
@@ -706,10 +706,10 @@ vector< map<string, int> > RemoveLineageCommand::getTaxons(string tax) {
 					newtaxon = taxon;
 					confidence = "0";
 				}
-				int con = 0;
+				float con = 0;
 				convert(confidence, con);
 				
-				map<string, int> temp;
+				map<string, float> temp;
 				temp[newtaxon] = con;
 				t.push_back(temp);
