+++ removelineagecommand.h
@@ -40,7 +40,7 @@ class RemoveLineageCommand : public Command {
 		int readList();
 		int readTax();	
 		string removeConfidences(string);
-		vector< map<string, int> > getTaxons(string);
+		vector< map<string, float> > getTaxons(string);
 };
 
 #endif
